Clean code!
Updated some logic to allow for the photo to be copied to an Archive directory first as opposed to with a cleanup loop. Also changed to boxcutter-fs.exe to see if it fixes the black screen issue. Took the opportunity to clean the code a little more (thanks Uncle Bob).