Updated ITGMania and related files!
Updated ITGMania to take advantage of the awesome Hades 2 theme included for SRPG7! Had to tweak my StartITGMania.au3 file to point to the new .exe. Good futureproofing making it portable, I simply dragged and dropped the Songs, Save and Noteskins folders into the new directory and called it a day.